player: Always enable previous button
Now that we can seek to the beginning of the current song with the
previous button, we don't need to lock the previous button down.
The logic to check if we can go back to a previous song is therefore
moved to previousTrackClicked. If there is no previous song, the
current song is simply replayed

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>